+++ src/main/java/ru/exchange/model/ExchangeRate.java
@@ -1,13 +1,15 @@
 package ru.exchange.model;
 
+import java.math.BigDecimal;
+
 public class ExchangeRate {
 
     private int id;
     private int baseCurrencyId;
     private int targetCurrencyId;
-    private double rate;
+    private BigDecimal rate;
 
-    public ExchangeRate(int baseCurrencyId, int targetCurrencyId, double rate) {
+    public ExchangeRate(int baseCurrencyId, int targetCurrencyId, BigDecimal rate) {
         this.baseCurrencyId = baseCurrencyId;
         this.targetCurrencyId = targetCurrencyId;
         this.rate = rate;
@@ -37,11 +39,11 @@ public class ExchangeRate {
         this.targetCurrencyId = targetCurrencyId;
     }
 
-    public double getRate() {
+    public BigDecimal getRate() {
         return rate;
     }
 
-    public void setRate(double rate) {
+    public void setRate(BigDecimal rate) {
         this.rate = rate;
     }
